Fix sphinx errors.
