+++ docs/api.rst
@@ -58,42 +58,3 @@ likelihood
 
     get_crit_func
     log_like
-
-
-Developer
----------
-
-config
-~~~~~~
-
-.. currentmodule:: respy.config
-
-.. autosummary::
-    :toctree: _generated/
-
-    INDEXER_DTYPE
-    INDEXER_INVALID_INDEX
-
-interpolate
-~~~~~~~~~~~
-
-.. currentmodule:: respy.interpolate
-
-.. autosummary::
-    :toctree: _generated/
-
-    interpolate
-    _kw_94_interpolation
-    _get_not_interpolated_indicator
-
-state space
-~~~~~~~~~~~
-
-.. currentmodule:: respy.state_space
-
-.. autosummary::
-    :toctree: _generated/
-
-    _BaseStateSpace
-    _SingleDimStateSpace
-    _MultiDimStateSpace
